add transformation

+++ src/main/kotlin/model/Point.kt
@@ -1,9 +1,9 @@
 package model
 
-data class Point(val x: Float, val y: Float) {
+data class Point(val x: Double, val y: Double) {
 
     companion object {
-        val Zero = Point(0f, 0f)
+        val Zero = Point(0.0, 0.0)
     }
 
 }
