+++ src/main/kotlin/model/Size.kt
@@ -1,3 +1,3 @@
 package model
 
-data class Size(val width: Float, val height: Float)
+data class Size(val width: Double, val height: Double)
