+++ src/test/kotlin/ParseSvgAttrs.kt
@@ -1,4 +1,4 @@
-import ext.nextFloat
+import ext.nextDouble
 import model.Svg
 import org.junit.Test
 import parser.SvgParser
@@ -11,10 +11,10 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg with valid width and height and viewBox`() {
-        val width = Random.nextFloat(255)
-        val height = Random.nextFloat(255)
-        val viewportWidth = Random.nextFloat(255)
-        val viewportHeight = Random.nextFloat(255)
+        val width = Random.nextDouble(255)
+        val height = Random.nextDouble(255)
+        val viewportWidth = Random.nextDouble(255)
+        val viewportHeight = Random.nextDouble(255)
         val a = Svg(width, height, viewportWidth, viewportHeight)
         val svg = svgParser.parseSvg(a.toString())
         assert(svg?.width == width && svg.height == height && svg.viewportWidth == viewportWidth && svg.viewportHeight == viewportHeight)
@@ -22,7 +22,7 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg without height`() {
-        val width = Random.nextFloat(255)
+        val width = Random.nextDouble(255)
         val a = "<svg width=\"$width\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.width == width && svg.height == width)
@@ -30,7 +30,7 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg without width`() {
-        val height = Random.nextFloat(255)
+        val height = Random.nextDouble(255)
         val a = "<svg height=\"$height\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.width == height && svg.height == height)
@@ -38,8 +38,8 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg without viewport`() {
-        val width = Random.nextFloat(255)
-        val height = Random.nextFloat(255)
+        val width = Random.nextDouble(255)
+        val height = Random.nextDouble(255)
         val a = "<svg width=\"$width\" height=\"$height\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.viewportWidth == width && svg.viewportHeight == height)
@@ -47,10 +47,10 @@ class ParseSvgAttrs {
 
     @Test
     fun `calc svg viewport width and height`() {
-        val l = Random.nextFloat(255)
-        val t = Random.nextFloat(255)
-        val r = l + Random.nextFloat(255) //just to make sure that right is bigger than left
-        val b = t + Random.nextFloat(255) //just to make sure that bottom is bigger than top
+        val l = Random.nextDouble(255)
+        val t = Random.nextDouble(255)
+        val r = l + Random.nextDouble(255) //just to make sure that right is bigger than left
+        val b = t + Random.nextDouble(255) //just to make sure that bottom is bigger than top
         val a = "<svg viewBox=\"$l $t $r $b\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.viewportWidth == r - l && svg.viewportHeight == b - t)
@@ -58,10 +58,10 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg without width and height`() {
-        val l = Random.nextFloat(255)
-        val t = Random.nextFloat(255)
-        val r = l + Random.nextFloat(255) //just to make sure that right is bigger than left
-        val b = t + Random.nextFloat(255) //just to make sure that bottom is bigger than top
+        val l = Random.nextDouble(255)
+        val t = Random.nextDouble(255)
+        val r = l + Random.nextDouble(255) //just to make sure that right is bigger than left
+        val b = t + Random.nextDouble(255) //just to make sure that bottom is bigger than top
         val a = "<svg viewBox=\"$l $t $r $b\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.width == r - l && svg.height == b - t)
@@ -69,13 +69,13 @@ class ParseSvgAttrs {
 
     @Test
     fun `calc svg viewport with left bigger that right and top bigger than bottom`() {
-        val r = Random.nextFloat(255)
-        val l = r + Random.nextFloat(255)
-        val b = Random.nextFloat(255)
-        val t = b + Random.nextFloat(255)
+        val r = Random.nextDouble(255)
+        val l = r + Random.nextDouble(255)
+        val b = Random.nextDouble(255)
+        val t = b + Random.nextDouble(255)
         val a = "<svg viewBox=\"$l $t $r $b\"></svg>"
         val svg = svgParser.parseSvg(a)
-        assert(svg?.viewportWidth == 0f && svg.viewportHeight == 0f)
+        assert(svg?.viewportWidth == 0.0 && svg.viewportHeight == 0.0)
     }
 
     @Test(expected = IllegalArgumentException::class)
@@ -92,8 +92,8 @@ class ParseSvgAttrs {
 
     @Test
     fun `parse svg with not valid viewBox and with width and height`() {
-        val width = Random.nextFloat(255)
-        val height = Random.nextFloat(255)
+        val width = Random.nextDouble(255)
+        val height = Random.nextDouble(255)
         val a = "<svg width=\"${width}\" height=\"${height}\" viewBox=\"4 5\"></svg>"
         val svg = svgParser.parseSvg(a)
         assert(svg?.viewportWidth == width && svg.viewportHeight == height)
